+++ src/pages/MapPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useLocalizedRoutes } from '../common/hooks/useLocalizedRoutes';
 import PageMeta from './components/PageMeta';
-import HankeMapContainer from '../domain/map/HankeMapContainer';
+import HankeMapContainer from '../domain/map/HankeMap';
 
 const MapPage: React.FC = () => {
   const { MAP } = useLocalizedRoutes();
